requirements.txt
fastapi
uvicorn
python-multipart
supabase
pillow
numpy
ultralytics